refactor: move GetDecisionDefinitionXmlTest tests to DecisionQueryTest
