+++ qa/integration-tests/src/test/java/io/camunda/it/client/DecisionDefinitionQueryTest.java
@@ -8,25 +8,33 @@
 package io.camunda.it.client;
 
 import static org.assertj.core.api.Assertions.assertThat;
+import static org.junit.jupiter.api.Assertions.assertThrows;
 
 import io.camunda.qa.util.cluster.TestStandaloneCamunda;
 import io.camunda.zeebe.client.ZeebeClient;
+import io.camunda.zeebe.client.api.command.ProblemException;
 import io.camunda.zeebe.client.api.response.Decision;
 import io.camunda.zeebe.client.api.response.DeploymentEvent;
 import io.camunda.zeebe.client.api.search.response.DecisionDefinition;
 import io.camunda.zeebe.client.impl.search.response.DecisionDefinitionImpl;
 import io.camunda.zeebe.qa.util.junit.ZeebeIntegration;
 import io.camunda.zeebe.qa.util.junit.ZeebeIntegration.TestZeebe;
+import java.io.IOException;
+import java.nio.file.Files;
+import java.nio.file.Paths;
 import java.time.Duration;
 import java.util.ArrayList;
 import java.util.Comparator;
 import java.util.List;
+import java.util.Objects;
+import java.util.Random;
+import java.util.concurrent.CompletionException;
 import org.awaitility.Awaitility;
 import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.Test;
 
 @ZeebeIntegration
-public class SearchDecisionDefinitionTest {
+public class DecisionDefinitionQueryTest {
   private static final List<Decision> DEPLOYED_DECISIONS = new ArrayList<>();
   private static ZeebeClient zeebeClient;
 
@@ -125,6 +133,39 @@ public class SearchDecisionDefinitionTest {
     assertThat(result.items().size()).isEqualTo(2);
     assertThat(result.items().get(0)).isEqualTo(toDecisionDefinition(decisionDefV2));
     assertThat(result.items().get(1)).isEqualTo(toDecisionDefinition(decisionDefV1));
+  }
+
+  @Test
+  void shouldGetDecisionDefinitionXml() throws IOException {
+    // when
+    final long decisionKey = DEPLOYED_DECISIONS.get(0).getDecisionKey();
+    final var result = zeebeClient.newDecisionDefinitionGetXmlRequest(decisionKey).send().join();
+
+    // then
+    final String expected =
+        Files.readString(
+            Paths.get(
+                Objects.requireNonNull(
+                        getClass().getClassLoader().getResource("decision/decision_model.dmn"))
+                    .getPath()));
+    assertThat(result).isNotEmpty();
+    assertThat(result).isEqualTo(expected);
+  }
+
+  @Test
+  void shouldReturn404ForNotFoundDecisionKey() {
+    // when
+    final long decisionKey = new Random().nextLong();
+    final var exception =
+        assertThrows(
+            CompletionException.class,
+            () -> zeebeClient.newDecisionDefinitionGetXmlRequest(decisionKey).send().join());
+    // then
+    assertThat(exception.getCause()).isInstanceOf(ProblemException.class);
+    final var problemException = (ProblemException) exception.getCause();
+    assertThat(problemException.code()).isEqualTo(404);
+    assertThat(problemException.details().getDetail())
+        .isEqualTo("DecisionDefinition with decisionKey=%d cannot be found".formatted(decisionKey));
   }
 
   private static void waitForDecisionsBeingExported() {
